需求[#18254]: AR-20200706-04-04 运维管理模块智能告警-告警记录查询接口开发
    [修改内容]:优化智能告警记录查询接口
    [描述说明]:
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/task/entity/TDefectInfoDetail.java
@@ -39,12 +39,12 @@ public class TDefectInfoDetail implements Serializable {
     private String confModeName;
 
     @ApiModelProperty(value = "设备id")
-    private Integer deviceId;
+    private Long deviceId;
     @ApiModelProperty(value = "缺陷主设备")
     private String deviceName;
 
     @ApiModelProperty(value = "缺陷点位id")
-    private Integer stdMeteId;
+    private Long stdMeteId;
     @ApiModelProperty(value = "缺陷点位名称")
     private String meteName;
